+++ Source/Lib/Common/Codec/EbMotionEstimation.c
@@ -8634,7 +8634,7 @@ static void quarter_pel_refinemnet_block(
         y_search_area_origin,  // [IN] search area origin in the vertical
                                // direction, used to point to reference samples
     EbAsm asm_type, uint32_t candidate_mv, uint32_t *p_best_sad,
-    uint32_t *p_best_mv, uint8_t is_frac_candidate) {
+    uint32_t *p_best_mv, uint16_t is_frac_candidate) {
     int16_t x_mv = _MVXT(candidate_mv);
     int16_t y_mv = _MVYT(candidate_mv);
     int16_t search_Index_x = ((x_mv + 2) >> 2) - x_search_area_origin;
